# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: CasaRoca System AI · API
  version: "1.0.0"
  description: |
    API del sistema eclesial de Casa Sobre la Roca: una central y 36 sedes.

    ESTE ARCHIVO SE GENERA. No lo edite a mano: corra `scripts/generar-openapi.js`.
    La verificacion de entrega compara lo generado con lo versionado.

    Autenticacion: token de portador (Bearer). El token dice que sesion DICE ser;
    quien decide es la base, que en cada peticion responde si esa sesion sigue viva.
    Por eso cerrar una sesion surte efecto en el instante.
servers:
  - url: https://api.casaroca.org
    description: produccion
  - url: https://api.staging.casaroca.org
    description: pruebas
  - url: http://127.0.0.1:3000
    description: desarrollo
components:
  securitySchemes:
    portador:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: boolean }
        mensaje: { type: string, description: "En castellano y con la accion siguiente. Nunca filtra la estructura interna." }
        peticionId: { type: string, description: "Identificador de traza. El usuario lo lee por telefono al soporte." }
security:
  - portador: []
paths:
...
  /api/v1/talento/antecedentes/:personaId:
    get:
      summary: "Los antecedentes de alguien"
      description: "Dice si esta apto para estar con menores y QUE le falta exactamente."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/talento/antecedentes:
    post:
      summary: "Registrar un antecedente"
      description: "La sede es la de la PERSONA (la base la pone y la del formulario se ignora): nadie sella antecedentes de otra sede. Solo lo registra quien alcanza a esa persona."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/talento/cargos:
    get:
      summary: "Los cargos"
      description: "Con su area y el nivel de dato que manejan."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/talento/contratos:
    get:
      summary: "Los contratos"
      description: "Avisa de los que vencen en menos de treinta dias."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/talento/voluntariados/:id/terminar:
    post:
      summary: "Terminar un voluntariado"
      description: "Exige motivo."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/talento/voluntariados:
    get:
      summary: "Los voluntariados"
      description: "⛔ Lo primero que se ve son los voluntarios ACTIVOS con menores y SIN antecedentes vigentes. No es una metrica: es una lista de salas que hay que cubrir antes del domingo."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    post:
      summary: "Registrar un voluntariado"
      description: "Persona, sede, ministerio y funcion."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...